Move collection normalization into content collection model
